# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API Assurances",
        "description": "Mon API de gestion des assurances",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur Local"
        }
    ],
    "paths": {
        "/api/assurances": {
            "get": {
                "tags": [
                    "Assurances"
                ],
                "summary": "Récupérer la liste des assurances",
                "operationId": "getAssurancesList",
                "responses": {
                    "200": {
                        "description": "Opération réussie"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Assurances"
                ],
                "summary": "Créer une nouvelle assurance",
                "operationId": "storeAssurance",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle",
                                    "montant",
                                    "bonus"
                                ],
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "Assurance Auto"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "example": 50000
                                    },
                                    "bonus": {
                                        "type": "number",
                                        "example": 10.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Assurance créée avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/assurances/{id}": {
            "get": {
                "tags": [
                    "Assurances"
                ],
                "summary": "Afficher les détails d'une assurance",
                "operationId": "getAssuranceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'assurance",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Assurances"
                ],
                "summary": "Mettre à jour une assurance existante",
                "operationId": "updateAssurance",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'assurance à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle",
                                    "montant",
                                    "bonus"
                                ],
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "Assurance Auto Modifiée"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "example": 60000
                                    },
                                    "bonus": {
                                        "type": "number",
                                        "example": 12.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Assurance mise à jour avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Assurances"
                ],
                "summary": "Supprimer une assurance",
                "operationId": "deleteAssurance",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'assurance à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assurance supprimée avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Se connecter pour obtenir un token",
                "operationId": "authLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie"
                    },
                    "401": {
                        "description": "Identifiants incorrects"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Inscrire un nouvel utilisateur",
                "operationId": "authRegister",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jean Dupont"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "jean.dupont@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Se déconnecter et détruire le token actuel",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Assurances",
            "description": "Assurances"
        },
        {
            "name": "Authentification",
            "description": "Authentification"
        }
    ]
}